otter-kds-v6/requirements.txt
# Core dependencies
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Web scraping and API
selenium==4.16.0
requests==2.31.0
httpx==0.25.2

# Data processing
pandas==2.1.4
openpyxl==3.1.2

# Async support
aiohttp==3.9.1
asyncio==3.4.3

# Database
sqlalchemy==2.0.23
alembic==1.13.0
supabase==2.3.0
postgrest==0.13.0

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.25.0
websockets==12.0
python-multipart==0.0.6

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0

# Code quality
black==23.12.0
ruff==0.1.8
mypy==1.7.1
types-requests==2.31.0.10

# Logging and monitoring
structlog==23.2.0

# CLI
click==8.1.7
rich==13.7.0